+++ Homework1/index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="style/style.css">
-    <link rel="stylesheet" href="style/normalize.css">
+    <!-- <link rel="stylesheet" href="style/style.css">
+    <link rel="stylesheet" href="style/normalize.css"> -->
     <title>Document</title>
 </head>
 <body>
@@ -39,6 +39,6 @@
     </main>
     
     <script defer src="https://use.fontawesome.com/releases/v5.7.2/js/all.js" integrity="sha384-0pzryjIRos8mFBWMzSSZApWtPl/5++eIfzYmTgBBmXYdhvxPc+XcFEk+zJwDgWbP" crossorigin="anonymous"></script>
-    <script src="js/main.js"></script>
+    <!-- <script src="js/app.js"></script> -->
 </body>
 </html>
