+++ Homework1/src/index.js
@@ -1,0 +1,5 @@
+//Точка входа 
+//Здесь осуществляются все импорты
+import app from './public/js/app.js'
+
+app()
